Updated to version 2
This is a complete re-write of the code. It does better seperation of the parsing logic from the object mapping code.
Has minimal tests for correctness (5% done)
Has a concrete object for the Context, but kinda mimics the offical go Context object (80% done)
The Router Handle logic doesn't rely on regex's anymore, so it should be more logical to follow
Added more value types that can be used for Argument and Optional objects
Laid in first stab of bubbling up the errors, but need to test more
Simple collision detection for handlers